[WIP] migrating Primer::ImageCrop to Primer::Alpha::ImageCrop (#1528)
