+++ .changeset/witty-shirts-exist.md
@@ -1,0 +1,5 @@
+---
+'@primer/view-components': patch
+---
+
+migrating Primer::ImageCrop to Primer::Alpha::ImageCrop
