+++ app/components/primer/alpha/image_crop.html.erb
