+++ app/components/primer/primer.ts
@@ -1,3 +1,4 @@
+import './alpha/image_crop'
 import './alpha/modal_dialog'
 import './alpha/nav_list'
 import './alpha/segmented_control'
@@ -6,7 +7,6 @@ import './alpha/tool_tip'
 import './beta/auto_complete/auto_complete'
 import './clipboard_copy'
 import './dropdown'
-import './image_crop'
 import './local_time'
 import './tab_container_component'
 import './time_ago_component'
