+++ static/arguments.json
@@ -358,6 +358,32 @@
     ]
   },
   {
+    "component": "ImageCrop",
+    "status": "alpha",
+    "source": "https://github.com/primer/view_components/tree/main/app/components/primer/alpha/image_crop.rb",
+    "lookbook": "https://primer.style/view-components/lookbook/inspect/primer/alpha/image_crop/default/",
+    "parameters": [
+      {
+        "name": "src",
+        "type": "String",
+        "default": "N/A",
+        "description": "The path of the image to crop."
+      },
+      {
+        "name": "rounded",
+        "type": "Boolean",
+        "default": "`true`",
+        "description": "If the crop mask should be a circle. Defaults to true."
+      },
+      {
+        "name": "system_arguments",
+        "type": "Hash",
+        "default": "N/A",
+        "description": "[System arguments](/system-arguments)"
+      }
+    ]
+  },
+  {
     "component": "Layout",
     "status": "alpha",
     "source": "https://github.com/primer/view_components/tree/main/app/components/primer/alpha/layout.rb",
@@ -469,7 +495,9 @@
   },
   {
     "component": "SegmentedControl",
+    "status": "alpha",
     "source": "https://github.com/primer/view_components/tree/main/app/components/primer/alpha/segmented_control.rb",
+    "lookbook": "https://primer.style/view-components/lookbook/inspect/primer/alpha/segmented_control/default/",
     "parameters": [
       {
         "name": "hide_labels",
@@ -1784,32 +1812,6 @@
         "type": "Boolean",
         "default": "`false`",
         "description": "Whether the button is in a [BorderBox](/components/beta/borderbox). If `true`, the button will have the `Box-btn-octicon` class."
-      },
-      {
-        "name": "system_arguments",
-        "type": "Hash",
-        "default": "N/A",
-        "description": "[System arguments](/system-arguments)"
-      }
-    ]
-  },
-  {
-    "component": "ImageCrop",
-    "status": "alpha",
-    "source": "https://github.com/primer/view_components/tree/main/app/components/primer/image_crop.rb",
-    "lookbook": "https://primer.style/view-components/lookbook/inspect/primer/image_crop/default/",
-    "parameters": [
-      {
-        "name": "src",
-        "type": "String",
-        "default": "N/A",
-        "description": "The path of the image to crop."
-      },
-      {
-        "name": "rounded",
-        "type": "Boolean",
-        "default": "`true`",
-        "description": "If the crop mask should be a circle. Defaults to true."
       },
       {
         "name": "system_arguments",
