+++ static/audited_at.json
@@ -12,6 +12,7 @@
   "Primer::Alpha::Dialog::Header": "",
   "Primer::Alpha::HiddenTextExpander": "",
   "Primer::Alpha::Image": "",
+  "Primer::Alpha::ImageCrop": "",
   "Primer::Alpha::Layout": "",
   "Primer::Alpha::Layout::Main": "",
   "Primer::Alpha::Layout::Sidebar": "",
