+++ static/statuses.json
@@ -12,6 +12,7 @@
   "Primer::Alpha::Dialog::Header": "alpha",
   "Primer::Alpha::HiddenTextExpander": "alpha",
   "Primer::Alpha::Image": "alpha",
+  "Primer::Alpha::ImageCrop": "alpha",
   "Primer::Alpha::Layout": "alpha",
   "Primer::Alpha::Layout::Main": "alpha",
   "Primer::Alpha::Layout::Sidebar": "alpha",
@@ -64,7 +65,7 @@
   "Primer::HellipButton": "alpha",
   "Primer::IconButton": "deprecated",
   "Primer::Image": "deprecated",
-  "Primer::ImageCrop": "alpha",
+  "Primer::ImageCrop": "deprecated",
   "Primer::LabelComponent": "beta",
   "Primer::LayoutComponent": "alpha",
   "Primer::LinkComponent": "beta",
